% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-utils.R
\name{apply_status}
\alias{apply_status}
\title{Give a status to a row.}
\usage{
apply_status(dat)
}
\arguments{
\item{dat}{Dataframe.}
}
\value{
A character vector containing statuses.
}
\description{
There can be 4 statuses: OK (not changed), IN REVIEW (changed but not validated),
ACCEPTED (reviewed) and REJECTED (reviewed but needs further changes).
}
