% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{assets}
\alias{assets}
\title{Assets}
\usage{
assets()
}
\description{
Includes all assets.
This is a convenience function that wraps
\link{serveAssets} and allows easily adding additional
remote dependencies (e.g.: CDN) should there be any.
}
\keyword{internal}
