% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-utils.R
\name{check_config}
\alias{check_config}
\title{Check user config yaml file}
\usage{
check_config(loader, session)
}
\arguments{
\item{loader}{Waiter loader R6 instance.}

\item{session}{Shiny session object.}
}
\value{
A waiter object
}
\description{
Run inside the init module.
}
