% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-utils.R
\name{check_if_user_logged}
\alias{check_if_user_logged}
\title{Check if we can find the connect user}
\usage{
check_if_user_logged(state, loader)
}
\arguments{
\item{state}{App state.}

\item{loader}{Screen loader to display feedback message
in case of error.}
}
\description{
Check if we can find the connect user
}
\keyword{internal}
