% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_get}
\alias{config_get}
\title{Get Value from Configuration}
\usage{
config_get(value, config = NULL)
}
\arguments{
\item{value}{The value to retrieve.}

\item{config}{The configuration (from \code{\link[=config_read]{config_read()}}),
if missing the configuration file is red.}
}
\description{
Retrieve a value from the configuration file.
}
