% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_read}
\alias{config_read}
\title{Read Config}
\usage{
config_read(path = getOption("app.config.path"))
}
\arguments{
\item{path}{Allow to pass custom pass to search for config file.}
}
\description{
Read Config
}
\keyword{internal}
