% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-utils.R
\name{create_table_cols}
\alias{create_table_cols}
\title{Create reactable columns config}
\usage{
create_table_cols(first_version, state)
}
\arguments{
\item{first_version}{Data for which to define the columns.}

\item{state}{App state.}
}
\value{
A list to pass to \link{edit_data_server}.
}
\description{
Initialise reactable column config for the data editor.
}
