% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-utils.R
\name{define_columns_diff}
\alias{define_columns_diff}
\title{Customize columns content}
\usage{
define_columns_diff(dat)
}
\arguments{
\item{dat}{Obtained from \link{get_first_version}.}
}
\value{
A list of options to pass to a reactable columns options.
}
\description{
Allows to setup a diff system so as to see whether data have changed.
Any change is depicted with red text next to the current value.
}
\note{
Only editable columns are considered.
}
\keyword{internal}
