% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-utils.R
\name{define_hidden_cols}
\alias{define_hidden_cols}
\title{Customize hidden columns}
\usage{
define_hidden_cols(dat)
}
\arguments{
\item{dat}{Data.}
}
\value{
A list of options to pass to a reactable columns options.
}
\description{
For each hidden column, indicate reactable to hide it
}
\keyword{internal}
