% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-utils.R
\name{find_cols_to_edit}
\alias{find_cols_to_edit}
\title{Find columns to edit}
\usage{
find_cols_to_edit(dat)
}
\arguments{
\item{dat}{Initial data without the editor columns.}
}
\value{
A vector of columns.
}
\description{
Used by \link{define_columns_diff}.
}
