% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-utils.R
\name{find_data_cols}
\alias{find_data_cols}
\title{Find the raw data cols}
\usage{
find_data_cols(dat)
}
\arguments{
\item{dat}{Initial data without the editor columns.}
}
\value{
A vector of columns.
}
\description{
Used by \link{define_columns_diff}.
}
