% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-utils.R
\name{find_factor_columns}
\alias{find_factor_columns}
\title{Find only factor type}
\usage{
find_factor_columns(dat)
}
\arguments{
\item{dat}{A dataframe containing columns metadata with name}
}
\description{
Find only factor type
}
\keyword{internal}
