% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-utils.R
\name{find_projects_to_lock}
\alias{find_projects_to_lock}
\title{Find project list to lock}
\usage{
find_projects_to_lock(dat, is_admin)
}
\arguments{
\item{dat}{Data to process.}

\item{is_admin}{Whether the current user belongs to an admin list.}
}
\value{
A boolean vector indicating whether the row should be locked. This
is useful later for JS.
}
\description{
Find project list to lock
}
