% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-utils.R
\name{get_data_version}
\alias{get_data_version}
\title{Get a specific version of pinned data}
\usage{
get_data_version(pin_name, board, versions, index)
}
\arguments{
\item{pin_name}{Pin name.}

\item{board}{Board.}

\item{versions}{Pins versions.}

\item{index}{Index to select}
}
\value{
The selected pin.
}
\description{
Get a specific version of pinned data
}
