% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-utils.R
\name{get_first_version}
\alias{get_first_version}
\title{Get first pin version}
\usage{
get_first_version(pin_name, board)
}
\arguments{
\item{pin_name}{Pin name.}

\item{board}{Board.}
}
\value{
A single pin.
}
\description{
Calls \link{get_data_version} on the very first pin version.
Useful to calculate the diff.
}
