% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-utils.R
\name{handle_validate_row}
\alias{handle_validate_row}
\title{Observer to handle row validation}
\usage{
handle_validate_row(action = c("accept", "reject"), state, board)
}
\arguments{
\item{action}{One of \code{accept} or \code{reject}.}

\item{state}{App state.}

\item{board}{Pins board.}
}
\description{
Rows can be rejected/accepted. This provides an observer to do so ...
}
