% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-utils.R
\name{is_user_admin}
\alias{is_user_admin}
\title{Find if current use is admin}
\usage{
is_user_admin(user, con)
}
\arguments{
\item{user}{App user. Given by state$user.}

\item{con}{Database pool.}
}
\value{
Boolean
}
\description{
Find if current use is admin
}
