% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets.R
\name{js-modules}
\alias{js-modules}
\alias{remove_modules}
\alias{get_modules}
\title{Module}
\usage{
remove_modules(files, modules)

get_modules(files, modules)
}
\arguments{
\item{files}{JavaScript files}

\item{modules}{JavaScript files names that require
the \code{type = module}.}
}
\description{
Retrieve and add modules from a vector of files.
}
\keyword{internal}
