% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lab}
\alias{lab}
\title{Laboratory data}
\format{
\subsection{\code{who}}{

A tibble with 59580 rows and 23 columns:
\describe{
\item{STUDYID}{Study ID}
\item{DOMAIN}{Data domain}
...
}
}
}
\source{
\url{https://github.com/cdisc-org/sdtm-adam-pilot-project/raw/master/updated-pilot-submission-package/900172/m5/datasets/cdiscpilot01/tabulations/sdtm/lb.xpt}
}
\usage{
lab
}
\description{
TBD
}
\keyword{datasets}
