% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-handlers.R
\name{leprechaun_handler_df}
\alias{leprechaun_handler_df}
\title{Input Dataframe}
\usage{
leprechaun_handler_df(data)
}
\arguments{
\item{data}{Input data received from WebSocket.}
}
\description{
Converts the input received from the WebSocket
to a data.frame.
}
\keyword{internal}
