% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-handlers.R
\name{leprechaun_handler_list}
\alias{leprechaun_handler_list}
\title{Input List}
\usage{
leprechaun_handler_list(data)
}
\arguments{
\item{data}{Input data received from WebSocket.}
}
\description{
Forces the input received from the WebSocket
to a list. This should really not be needed as
it is handled like so by default.
}
\keyword{internal}
