% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-utils.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare data}
\usage{
prepare_data(con, dat = lab, overwrite = FALSE)
}
\arguments{
\item{con}{Database pool.}

\item{dat}{Input data. Must be a dataframe or tibble.}

\item{overwrite}{Whether to reset the existing table. Default
to FALSE.}
}
\value{
Save new data in the provided database table.
}
\description{
Add extra columns to a given dataset needed by the editor app.
}
