% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_refresh_data.R
\name{refresh_dataUI}
\alias{refresh_dataUI}
\title{refresh_data UI}
\usage{
refresh_dataUI(id)
}
\arguments{
\item{id}{Unique id for module instance.}
}
\description{
refresh_data UI
}
\keyword{internal}
