% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_reset.R
\name{reset_server}
\alias{reset_server}
\title{reset Server}
\usage{
reset_server(id, con, screen_loader)
}
\arguments{
\item{id}{Unique id for module instance.}

\item{con}{Database pool.}

\item{screen_loader}{Waiter R6 instance.}
}
\description{
reset Server
}
\keyword{internal}
