% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-utils.R
\name{restore_col_type}
\alias{restore_col_type}
\title{Check column type compared to metadata}
\usage{
restore_col_type(x)
}
\arguments{
\item{x}{Column name.}
}
\description{
If type does not match, the right type is restored
from the metadata.
}
\keyword{internal}
