% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Run}
\usage{
run(board, ...)
}
\arguments{
\item{board}{Board to pin data.}

\item{...}{Additional parameters to pass to \link[shiny:shinyApp]{shiny::shinyApp}.}
}
\description{
Run application
}
