% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assets.R
\name{serveAssets}
\alias{serveAssets}
\title{Dependencies}
\usage{
serveAssets(modules = NULL)
}
\arguments{
\item{modules}{JavaScript files names that require
the \code{type = module}.}
}
\description{
Dependencies
}
\keyword{internal}
