% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-utils.R
\name{split_data_cols}
\alias{split_data_cols}
\title{Split data cols}
\usage{
split_data_cols(dat)
}
\arguments{
\item{dat}{Data to process.}
}
\value{
A list with 2 entries.
}
\description{
Cols are split into editable cols and showable cols.
}
