% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_unlock_row.R
\name{unlock_row_server}
\alias{unlock_row_server}
\title{unlock_row Server}
\usage{
unlock_row_server(id, trigger, dat, row_index, con)
}
\arguments{
\item{id}{Unique id for module instance.}

\item{trigger}{Reactive trigger.}

\item{dat}{App data.}

\item{row_index}{Row to edit.}

\item{con}{Database pool.}
}
\description{
unlock_row Server
}
\keyword{internal}
