% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_validate_row.R
\name{validate_row_server}
\alias{validate_row_server}
\title{validate_row Server}
\usage{
validate_row_server(id, trigger, dat, con)
}
\arguments{
\item{id}{Unique id for module instance.}

\item{trigger}{Reactive trigger.}

\item{dat}{Data to validate.}

\item{con}{Database pool.}
}
\description{
validate_row Server
}
\keyword{internal}
