% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_validate_row.R
\name{validate_rowUI}
\alias{validate_rowUI}
\title{validate_row UI}
\usage{
validate_rowUI(id)
}
\arguments{
\item{id}{Unique id for module instance.}
}
\description{
validate_row UI
}
\keyword{internal}
