% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-utils.R
\name{whoami}
\alias{whoami}
\title{Get current user
Must be called from the shiny server function.
On Posit connect, get \code{session$user}. Locally will try to run \code{whoami}.
If not found hardcode the current user to \code{test-user}.}
\usage{
whoami(session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{session}{Shiny session object.}
}
\value{
The current user.
}
\description{
Get current user
Must be called from the shiny server function.
On Posit connect, get \code{session$user}. Locally will try to run \code{whoami}.
If not found hardcode the current user to \code{test-user}.
}
