% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-utils.R
\name{with_tooltip}
\alias{with_tooltip}
\title{Adds tooltip to table header}
\usage{
with_tooltip(value, tooltip)
}
\arguments{
\item{value}{Text.}

\item{tooltip}{tooltip.}
}
\value{
Adds tooltip to a given tag.
}
\description{
Adds tooltip to table header
}
