% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_allow_save.R
\name{allow_save_server}
\alias{allow_save_server}
\title{allow_save Server}
\usage{
allow_save_server(id, trigger, state)
}
\arguments{
\item{id}{Unique id for module instance.}

\item{trigger}{Reactive trigger.}

\item{state}{App state.}
}
\description{
allow_save Server
}
\keyword{internal}
