% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_init.R
\name{init_server}
\alias{init_server}
\title{init Server}
\usage{
init_server(id, con, state, screen_loader)
}
\arguments{
\item{id}{Unique id for module instance.}

\item{con}{Database pool.}

\item{state}{App state.}

\item{screen_loader}{Waiter R6 instance.}
}
\description{
init Server
}
\keyword{internal}
