% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_init.R
\name{init_server}
\alias{init_server}
\title{init Server}
\usage{
init_server(id, state, screen_loader, first_version, input_data)
}
\arguments{
\item{id}{Unique id for module instance.}

\item{state}{App state.}

\item{screen_loader}{Waiter R6 instance.}

\item{first_version}{Data to send to JS. First pin version is fine.}

\item{input_data}{Input data provided by a reactive pin.}
}
\description{
init Server
}
\keyword{internal}
