% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_lock_row.R
\name{lock_row_server}
\alias{lock_row_server}
\title{lock_row Server}
\usage{
lock_row_server(id, trigger, state, board, screen_loader)
}
\arguments{
\item{id}{Unique id for module instance.}

\item{trigger}{Reactive trigger.}

\item{state}{App state.}

\item{board}{Where to save data. Pins.}

\item{screen_loader}{Waiter R6 instance.}
}
\description{
lock_row Server
}
\keyword{internal}
