% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-utils.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare data}
\usage{
prepare_data(dat, board, pin_name)
}
\arguments{
\item{dat}{Input data. Must be a dataframe or tibble.}

\item{board}{Board to save data.}

\item{pin_name}{Pin name.}
}
\value{
Save new data to a pin.
}
\description{
Add extra columns to given dataset needed by the editor app.
}
