% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_refresh_data.R
\name{refresh_data_server}
\alias{refresh_data_server}
\title{refresh_data Server}
\usage{
refresh_data_server(id, state, current_page, con, screen_loader)
}
\arguments{
\item{id}{Unique id for module instance.}

\item{state}{App state.}

\item{current_page}{Currently active table page.}

\item{con}{Pool DB connection.}

\item{screen_loader}{Waiter R6 instance.}
}
\description{
refresh_data Server
}
\keyword{internal}
