% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_reset.R
\name{reset_server}
\alias{reset_server}
\title{reset Server}
\usage{
reset_server(id, data_reset, board)
}
\arguments{
\item{id}{Unique id for module instance.}

\item{data_reset}{Data to treat.}

\item{board}{Where to store data.}
}
\description{
reset Server
}
\keyword{internal}
