% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{run}
\alias{run}
\title{Run}
\usage{
run(pool, ...)
}
\arguments{
\item{pool}{Database pool.}

\item{...}{Additional parameters to pass to \link[shiny:shinyApp]{shiny::shinyApp}.}
}
\description{
Run application
}
