% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_save_data.R
\name{save_data_server}
\alias{save_data_server}
\title{save_data Server}
\usage{
save_data_server(id, trigger, new_data, row_index, board, cache)
}
\arguments{
\item{id}{Unique id for module instance.}

\item{trigger}{Reactive trigger.}

\item{new_data}{New data from the table.}

\item{row_index}{Row to edit.}

\item{board}{Where to save data.}

\item{cache}{Cached data.}
}
\description{
save_data Server
}
\keyword{internal}
