% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-utils.R
\name{setup_pool}
\alias{setup_pool}
\title{Setup database pool}
\usage{
setup_pool(driver)
}
\arguments{
\item{driver}{DB driver.}
}
\value{
A database pool
}
\description{
Establish a database connection with the
provided driver.
}
